requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
pymupdf==1.24.9
